asta-core/requirements.txt
# Core Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0

# Environment & Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Code quality helpers (Optional but highly recommended for development)
black==23.11.0  # Auto-formatting
isort==5.12.0   # Sorts imports
